Move Renderable in the root namespace.
Being renderable has nothing to do with a renderer in this case
(although, theoretically they _can_).
Renderable is ok for basic rendering of objects, to enforce their
classes to implement the __toString() method but rendering a Quote,
becomes more complex when we have to render it in different formats and
so on... Therefor it seems best to keep Renderable for simple things and
expand Render/ namespace to go beyond __toString() and to make renderers
depend on Attributable, but not Renderable.